feat: replaced the fetch interval value adjuster with a slider

+++ GitBar/DefaultsExtensions.swift
@@ -10,7 +10,7 @@ import SwiftUI
 
 extension Defaults.Keys {
 	static let githubUsername = Key<String>("githubUsername", default: "guppy57", iCloud: true)
-	static let fetchInterval = Key<Double>("fetchInterval", default: 5.0, iCloud: true)
+	static let fetchInterval = Key<Int>("fetchInterval", default: 5, iCloud: true)
 	static let githubToken = Key<String>("githubToken", default: "", iCloud: false)
 	static let rollingFetchDays = Key<Double>("rollingFetchDays", default: 1.0, iCloud: true)
 }
